Refactor: server api, client api 수정, 리팩토링
