--- src/container/interviewResult/services/files/index.ts
@@ -1,10 +1,0 @@
-import fetchAPI from '@/container/questions/libs/fetchAPI';
-
-import { IGetInterviewVideoUrl } from '../../types/files';
-
-export const getInterviewVideoUrl = (
-  videoId: number,
-): Promise<IGetInterviewVideoUrl> => {
-  const url = `api/files/download-url/${videoId}`;
-  return fetchAPI(url);
-};
