--- src/container/interviewResult/types/files/index.ts
@@ -1,6 +1,0 @@
-export interface IGetInterviewVideoUrl {
-  message: string;
-  data: {
-    downloadUrl: string;
-  };
-}
