--- src/container/questions/libs/fetchAPI/types.ts
@@ -1,1 +1,0 @@
-export type TMethod = 'get' | 'post' | 'delete' | 'put' | 'patch';
