+++ src/libs/types/params.ts
@@ -1,5 +1,18 @@
+// questions
 export interface IGetQuestionByIdParams {
   questionId: number;
   page: number;
   size: number;
 }
+
+// interviews
+
+// gpt
+
+// categories
+
+// auth
+
+// answers
+
+// files
