+++ src/libs/types/payload.ts
@@ -1,0 +1,21 @@
+// questions
+export interface IPatchQuestionPayload {
+  content: string;
+  categoryIds: number[];
+}
+
+// interviews
+export interface IPatchInterviewVisibilityPayload {
+  answerId: number;
+  visibility: 'PUBLIC' | 'PRIVATE';
+}
+
+// gpt
+
+// categories
+
+// auth
+
+// answers
+
+// files
