+++ src/types/answers.ts
@@ -1,0 +1,4 @@
+// clickAnswerHeart
+export interface IAnswerHeart {
+  isHearted: boolean;
+}
