+++ src/types/categories.ts
@@ -1,0 +1,5 @@
+// getCategories
+export interface ICategory {
+  id: number;
+  name: string;
+}
