+++ src/types/files.ts
@@ -1,0 +1,4 @@
+// getInterviewVideoUrl
+export interface IDownLoadUrl {
+  downloadUrl: string;
+}
